송금액은 소수점을 사용하지 않으므로 데이터 타입을 정수타입으로 변경

+++ src/main/java/dev/jihun/domain/exchangerate/dto/ReceivedAmountCalculationRequest.java
@@ -13,6 +13,6 @@ public class ReceivedAmountCalculationRequest {
 
     @Min(value = 1, message = "송금액은 1 USD 이상이여야 합니다.")
     @Max(value = 10000, message = "송금액은 10000 미만이여야 합니다.")
-    private double remittanceAmount;
+    private int remittanceAmount;
 
 }
